# ex1/swagger.yaml
openapi: 3.0.0
info:
  title: Board Games API
  description: API for managing a board games dataset
  version: 1.0.0
servers:
  - url: http://localhost:17000
paths:
  /jogos:
    get:
      summary: Get all games or games by publisher
      parameters:
        - name: editora
          in: query
          schema:
            type: string
          description: Publisher name (filter)
      responses:
        '200':
          description: A list of games
    post:
      summary: Add a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jogo'
      responses:
        '201':
          description: Game created
        '500':
          description: Error creating game
  /jogos/{id}:
    get:
      summary: Get a game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game details
    put:
      summary: Update a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jogo'
      responses:
        '200':
          description: Game updated
    delete:
      summary: Delete a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game deleted
  /autores:
    get:
      summary: Get all authors and their games
      responses:
        '200':
          description: List of authors
  /categorias:
    get:
      summary: Get all categories and their games
      responses:
        '200':
          description: List of categories

components:
  schemas:
    Jogo:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        year:
          type: integer
        category:
          type: string
        minPlayers:
          type: integer
        maxPlayers:
          type: integer
        playingTimeMinutes:
          type: integer
        descriptionEN:
          type: string
        autores:
          type: array
          items:
            $ref: '#/components/schemas/Autor'
        editoras:
          type: array
          items:
            $ref: '#/components/schemas/Editora'
        mecanicas:
          type: array
          items:
            $ref: '#/components/schemas/Mecanica'
        premios:
          type: array
          items:
            $ref: '#/components/schemas/Premio'
    Autor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    Editora:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        country:
          type: string
    Mecanica:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    Premio:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        year:
          type: integer